requirements.txt
asgiref==3.3.1
black==21.12b0
certifi==2020.12.5
chardet==4.0.0
click==8.0.3
dj-database-url==0.5.0
Django==3.1.6
django-appconf==1.0.5
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-imagekit==4.1.0
django-js-asset==1.2.2
django-rosetta==0.9.5
django-taggit==2.0.0
idna==2.10
mypy-extensions==0.4.3
pathspec==0.9.0
pilkit==2.0
Pillow==9.0.0
platformdirs==2.4.1
polib==1.1.0
psycopg2-binary==2.9.3
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.3
